requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
requests>=2.31.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.18
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
slowapi>=0.1.9
boto3>=1.34.0
jinja2>=3.1.4
email-validator>=2.1.1
